backend/requirements.txt
# ============================================================================
#  BBOCR Backend Dependencies
# ============================================================================
#
#  주의: paddlepaddle-gpu는 CUDA 버전에 따라 별도 설치해야 합니다.
#        이 파일에는 포함하지 않았으므로 README.md의 설치 가이드를 참고하세요.
#
# ============================================================================

# ── FastAPI 서버 ──
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ── PaddleOCR + PaddleX (레이아웃 감지) ──
paddleocr>=3.1.0
paddlex>=3.1.0

# ── 이미지 처리 ──
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# ── PDF 처리 ──
PyMuPDF>=1.23.0
pdf2image>=1.16.0
reportlab>=4.0.0
PyPDF2>=3.0.0

# ── 데이터 모델 ──
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── 유틸리티 ──
PyYAML>=6.0
aiofiles>=23.0.0
python-dateutil>=2.8.0

# ── DB ──
sqlalchemy>=2.0.0

# ── 기타 ──
scikit-learn>=1.3.0
